Refactor file structure for current and future versioning (#42)
* Created a v2-0 version

Duplicated the MVP v0.1 folder to create a v2 version for the forthcoming rounds of iterations

Signed-off-by: Andrew Scrivener <andrew.scrivener@cps.gov.uk>

* Refactored the existing v0.1 directory name

Changed from
/version-0/0_1-housekeeping/

To
/version-0/0-housekeeping_0-1

As I’ve been asked to keep historic versions for reference so this should be easy to understand for future devs.

Signed-off-by: Andrew Scrivener <andrew.scrivener@cps.gov.uk>

* Fixed typo

Signed-off-by: Andrew Scrivener <andrew.scrivener@cps.gov.uk>

* Project name is “Case material and communications”

Changed following discussions within the team. Not “Housekeeping” or “Casework app” or “Case management” or “Merge”.

Signed-off-by: Andrew Scrivener <andrew.scrivener@cps.gov.uk>

* Crown Prosecution Service abbreviated as CPS

Crown Prosecution Service abbreviated as CPS to align with the conventions of other related services.

Signed-off-by: Andrew Scrivener <andrew.scrivener@cps.gov.uk>

* MVP version 0-1 changed to version 1

Refactored version 0-1 within the version 0 folder to be version 1 in a new folder now that it’s been confirmed as the Merge MVP.

Signed-off-by: Andrew Scrivener <andrew.scrivener@cps.gov.uk>

* Version 2-0 changed to version x

Placeholder created for a new ‘future’ version to allow us to look at possible future enhancements without being tied to a specific release number.

Signed-off-by: Andrew Scrivener <andrew.scrivener@cps.gov.uk>

* Removed the now unnecessary _v2 suffix

As part of moving version specific includes into sub folders the _v2 suffix is unnecessary.

Signed-off-by: Andrew Scrivener <andrew.scrivener@cps.gov.uk>

* Document paths updated

Signed-off-by: Andrew Scrivener <andrew.scrivener@cps.gov.uk>

* Project title changed

Signed-off-by: Andrew Scrivener <andrew.scrivener@cps.gov.uk>

* Renamed version folder to lowercase x

Signed-off-by: Andrew Scrivener <andrew.scrivener@cps.gov.uk>

* Moved version specific includes to subfolders

Restructuring to move baseline includes to a ‘from-version-1-0’ sub-folder. Any future version specific variations can be added and tested from their folders to keep the file/folder structure cleaner hopefully.

Signed-off-by: Andrew Scrivener <andrew.scrivener@cps.gov.uk>

* Removed superfluous ‘housekeeping’ in filenames

Now that it’s a combined housekeeping and casework application, there’s no need to include this in filenames or directories moving forward.

Signed-off-by: Andrew Scrivener <andrew.scrivener@cps.gov.uk>

* Updated references to use from v1 includes

Future changes will use version one (MVP) as the baseline with overrides if necessary.

Signed-off-by: Andrew Scrivener <andrew.scrivener@cps.gov.uk>

* Untrack DS_Store files

Signed-off-by: Andrew Scrivener <andrew.scrivener@cps.gov.uk>

* MVP banner variant added

Created a variant of the V7 banner to use as the MVP baseline and imported it into the v1 and vx pages.

Signed-off-by: Andrew Scrivener <andrew.scrivener@cps.gov.uk>

* File paths updated for editing statements and exhibits

Signed-off-by: Andrew Scrivener <andrew.scrivener@cps.gov.uk>

* Paths updated for search page

Signed-off-by: Andrew Scrivener <andrew.scrivener@cps.gov.uk>

* Duplicated includes across for vx

We need to adopt a more robust method of versioning these files instead of duplicating them, probably using Express routing or variables.

Signed-off-by: Andrew Scrivener <andrew.scrivener@cps.gov.uk>

* Updated paths and minor content tweaks

Signed-off-by: Andrew Scrivener <andrew.scrivener@cps.gov.uk>

* Updated relative include paths

Signed-off-by: Andrew Scrivener <andrew.scrivener@cps.gov.uk>

* Paths updated

Signed-off-by: Andrew Scrivener <andrew.scrivener@cps.gov.uk>

---------

Signed-off-by: Andrew Scrivener <andrew.scrivener@cps.gov.uk>